+++ src/main/java/beechat/Beechat.java
@@ -1,5 +1,11 @@
 package beechat;
 
+import beechat.command.Command;
+import beechat.task.TaskList;
+import beechat.util.Parser;
+import beechat.util.Storage;
+import beechat.util.Ui;
+
 import java.lang.String;
 import java.io.IOException;
