+++ src/main/java/beechat/command/DeleteCommand.java
@@ -1,4 +1,9 @@
-package beechat;
+package beechat.command;
+
+import beechat.util.Storage;
+import beechat.task.Task;
+import beechat.task.TaskList;
+import beechat.util.Ui;
 
 import java.io.IOException;
