+++ src/main/java/beechat/exception/BeeException.java
@@ -1,4 +1,4 @@
-package beechat;
+package beechat.exception;
 
 public class BeeException extends RuntimeException {
     public BeeException(String message) {
