+++ src/main/java/beechat/task/Task.java
@@ -1,6 +1,6 @@
-package beechat;
+package beechat.task;
 
-abstract class Task {
+public abstract class Task {
     protected String description;
     protected boolean isDone = false;
