+++ src/main/java/beechat/task/TodoTask.java
@@ -1,6 +1,6 @@
-package beechat;
+package beechat.task;
 
-class TodoTask extends Task {
+public class TodoTask extends Task {
     public TodoTask(String description) {
         super(description);
     }
